+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.0.6
+
+* minore bug fixed
+
+## 0.0.5
+
+* Un-initialized loader added.
+
+## 0.0.4
+
+* Pause & Resume options added.
+
+
+## 0.0.3
+
+* AutoPlay configuration added.
+
+## 0.0.2
+
+* Overlay not hiding with controls problem fixed.
+
 ## 0.0.1
 
 * initial release.